Add files via upload
This file makes it easier to see the training data because it is organized in a table in Excel. I think that we should include this file just for visualization purposes, because viewing the raw data as we have it right now is very messy and unclear. Moreover, if the professor or TA's want to view the data, this will make it somewhat easier. 

I am not sure if this is what Yuchen is working on, but what do you think of including this file somewhere in the repository?